More heavy duty refactoring
