+++ angular-client/src/app/components/participants/add-participant/add-participant.component.ts
