+++ angular-client/src/app/components/participants/view-participants/view-participants.component.css
