+++ angular-client/src/app/components/participants/view-participants/view-participants.component.html
@@ -1,0 +1,20 @@
+<br>
+<div *ngFor="let participant of profiles">
+<mat-card class="example-card">
+  <mat-card-header>
+    <div mat-card-avatar class="example-header-image"></div>
+    <mat-card-title><b>Name: {{participant.name}}</b></mat-card-title>
+    <mat-card-subtitle>Username: {{participant._id}}</mat-card-subtitle>
+  </mat-card-header>
+  <mat-card-content>
+   <div *ngIf="participant.email !== ''"><b> Email: </b> {{participant.email}} <br></div>
+   <div *ngIf="participant.telephone !== '' "><b> Phone number: </b> {{participant.telephone}} <br></div>
+   <div *ngIf="participant.socialmedia !== undefined"><b> Social media: </b> {{participant.socialmedia.service}}  {{participant.socialmedia.username}}<br></div>
+   <div *ngIf="participant.address !== '' "><b> Address: </b> {{participant.address}}<br></div>
+  </mat-card-content>
+  <mat-card-actions>
+      <a routerLink="/participant-profile/{{participant._id}}"><button mat-raised-button>View Participant</button></a>
+    <button mat-raised-button (click)="delete(participant._id)">Delete Participant</button>
+  </mat-card-actions>
+</mat-card><br>
+</div>
