--- angular-client/src/app/components/profiles/profiles.component.html
@@ -1,20 +1,0 @@
-<br>
-<div *ngFor="let profile of profiles">
-<mat-card class="example-card">
-  <mat-card-header>
-    <div mat-card-avatar class="example-header-image"></div>
-    <mat-card-title><b>Name: {{profile.name}}</b></mat-card-title>
-    <mat-card-subtitle>Username: {{profile._id}}</mat-card-subtitle>
-  </mat-card-header>
-  <mat-card-content>
-   <div *ngIf="profile.email !== ''"><b> Email: </b> {{profile.email}} <br></div>
-   <div *ngIf="profile.telephone !== '' "><b> Phone number: </b> {{profile.telephone}} <br></div>
-   <div *ngIf="profile.socialmedia !== undefined"><b> Social media: </b> {{profile.socialmedia.service}}  {{profile.socialmedia.username}}<br></div>
-   <div *ngIf="profile.address !== '' "><b> Address: </b> {{profile.address}}<br></div>
-  </mat-card-content>
-  <mat-card-actions>
-      <a routerLink="/participant-profile/{{profile._id}}"><button mat-raised-button>View Profile</button></a>
-    <button mat-raised-button (click)="delete(profile._id)">Delete Profile</button>
-  </mat-card-actions>
-</mat-card><br>
-</div>
